terminado, falta revision final de todo

+++ Proyecto-2/ordenes-service/src/main/resources/db/migration/V2__crear_tabla_detalle_orden.sql
@@ -1,7 +1,7 @@
 CREATE TABLE detalle_orden (
               id_detalle BIGINT AUTO_INCREMENT PRIMARY KEY,
-              cantidad INT NOT NULL,
-              subtotal DOUBLE NOT NULL,
+              cantidad integer NOT NULL,
+              subtotal integer NOT NULL,
               id_orden BIGINT NOT NULL,
               id_producto BIGINT NOT NULL,
